third_party/blink/web_tests/external/wpt/scroll-animations/constructor-expected.txt

This is a testharness.js-based test.
Found 67 tests; 65 PASS, 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS A ScrollTimeline can be created with a scrollSource
PASS A ScrollTimeline can be created with a non-scrolling scrollSource
PASS A ScrollTimeline created with a null scrollSource should have no scrollSource
PASS A ScrollTimeline created without a scrollSource should use the document.scrollingElement
PASS A ScrollTimeline created with the default orientation should default to 'block'
PASS 'block' is a valid orientation value
PASS 'inline' is a valid orientation value
PASS 'horizontal' is a valid orientation value
PASS 'vertical' is a valid orientation value
PASS Creating a ScrollTimeline with an invalid orientation value should throw
PASS A ScrollTimeline created with the default startScrollOffset should default to CSSKeywordValue(auto)
PASS A ScrollTimeline created with the default endScrollOffset should default to CSSKeywordValue(auto)
PASS CSSKeywordValue(auto) is a valid scroll offset value
PASS CSSUnitValue(0px) is a valid scroll offset value
PASS CSSMathSum(calc(100% + -80px)) is a valid scroll offset value
PASS 'auto' is a valid scroll offset value
PASS 'em' is a valid scroll offset unit
PASS 'ex' is a valid scroll offset unit
PASS 'ch' is a valid scroll offset unit
PASS 'rem' is a valid scroll offset unit
PASS 'vw' is a valid scroll offset unit
PASS 'vh' is a valid scroll offset unit
PASS 'vmin' is a valid scroll offset unit
PASS 'vmax' is a valid scroll offset unit
PASS 'cm' is a valid scroll offset unit
PASS 'mm' is a valid scroll offset unit
PASS 'q' is a valid scroll offset unit
PASS 'in' is a valid scroll offset unit
PASS 'pc' is a valid scroll offset unit
PASS 'pt' is a valid scroll offset unit
PASS 'px' is a valid scroll offset unit
PASS '%' is a valid scroll offset unit
PASS '' is an invalid scroll offset value
PASS 'calc(360deg / 4)' is an invalid scroll offset value
PASS 'left' is an invalid scroll offset value
PASS '#ff0000' is an invalid scroll offset value
PASS 'rgb(0, 128, 0)' is an invalid scroll offset value
PASS 'url("http://www.example.com/pinkish.gif")' is an invalid scroll offset value
PASS 'this_is_garbage' is an invalid scroll offset value
PASS CSSUnitValue(0) is an invalid scroll offset value
PASS '100px 5%' is an invalid scroll offset value
PASS '0' is an invalid scroll offset value
PASS '10px' is an invalid scroll offset value
PASS '10%' is an invalid scroll offset value
PASS 'calc(100% - 80px)' is an invalid scroll offset value
PASS 'deg' is an invalid scroll offset unit
PASS 's' is an invalid scroll offset unit
PASS 'Hz' is an invalid scroll offset unit
PASS 'dpi' is an invalid scroll offset unit
PASS '{"target":{}}' is a valid scroll offset value
PASS '{"target":{},"threshold":0}' is a valid scroll offset value
PASS '{"target":{},"threshold":0.5}' is a valid scroll offset value
PASS '{"target":{},"threshold":1}' is a valid scroll offset value
PASS '{}' is an invalid scroll offset value
PASS '{"target":{},"threshold":"test"}' is an invalid scroll offset value
PASS '{"target":{},"threshold":2}' is an invalid scroll offset value
PASS '{"target":{},"threshold":-0.2}' is an invalid scroll offset value
FAIL A ScrollTimeline created with the default timeRange should default to 'auto' Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported
FAIL 'auto' is a valid timeRange value Failed to construct 'ScrollTimeline': 'auto' value for timeRange not yet supported
PASS '0' is a valid timeRange value
PASS '-100' is a valid timeRange value
PASS '100' is a valid timeRange value
PASS '1234.5678' is a valid timeRange value
PASS 'invalid' is an invalid timeRange value
PASS 'Infinity' is an invalid timeRange value
PASS '-Infinity' is an invalid timeRange value
PASS 'NaN' is an invalid timeRange value
Harness: the test ran to completion.
